fix and pass all the data and get the response
